requirements.txt
django==4.2.5
pillow==10.0.1
django-environ==0.11.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
djangorestframework-simplejwt==5.3.0
factory_boy